Revert "Move styles to styles.ts"
