+++ client/src/app.css
@@ -1,0 +1,6 @@
+.App {
+  font-family: "Inter", -apple-system, Arial, Sans-Serif;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 20px;
+}
