--- client/src/components/Alert/styles.ts
@@ -1,9 +1,0 @@
-import { makeStyles, Theme } from "@material-ui/core/styles";
-
-const useStyles = makeStyles((theme: Theme) => {
-  return {
-    root: {},
-  };
-});
-
-export default useStyles;
