+++ client/src/components/CatalogCard/CatalogCreator.tsx
@@ -1,3 +1,5 @@
+import { makeStyles } from "@material-ui/core/styles";
+
 import { CardActionArea, Typography, Card } from "@material-ui/core/";
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
@@ -6,8 +8,31 @@ import { useState } from "react";
 import FormCreator from "../FormCreator/FormCreator";
 import createCatalog from "../../api/createCatalog";
 
-//STYLES
-import useStyles from "./styles";
+const useStyles = makeStyles(() => ({
+  root: {
+    height: 325,
+    borderRadius: "8px",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    padding: "16px",
+  },
+  cardContent: {
+    height: "100%",
+    border: "1px dashed #6A5DF9",
+    borderRadius: "4px",
+    display: "flex",
+    flexDirection: "column",
+    alignItems: "center",
+    justifyContent: "center",
+    color: "#6A5DF9",
+    textAlign: "center",
+    padding: "16px",
+  },
+  icon: {
+    fontWeight: 800,
+  },
+}));
 
 export default function CatalogCreator({ className }: any) {
   const classes = useStyles();
@@ -15,12 +40,9 @@ export default function CatalogCreator({ className }: any) {
   const handleModal = () => setOpen((prev) => !prev);
 
   return (
-    <>
-      <Card className={classes.cardCreator}>
-        <CardActionArea
-          className={classes.creatorContent}
-          onClick={handleModal}
-        >
+    <div className={className}>
+      <Card className={classes.root}>
+        <CardActionArea className={classes.cardContent} onClick={handleModal}>
           <FontAwesomeIcon className={classes.icon} size="2xl" icon={faPlus} />
           <Typography variant="body2">Create a new catalog</Typography>
         </CardActionArea>
@@ -34,6 +56,6 @@ export default function CatalogCreator({ className }: any) {
           acceptBtnName="Create"
         />
       )}
-    </>
+    </div>
   );
 }
