+++ client/src/components/CustomTabs.tsx
@@ -1,10 +1,16 @@
 import React from "react";
 import { Paper, Tabs, Tab } from "@material-ui/core";
-import useStyles from "./styles";
+import { makeStyles } from "@material-ui/core/styles";
 
 interface TpropsArray {
   tabValues: { columnName: string; content: any; disabled?: boolean }[];
 }
+
+const useStyles = makeStyles({
+  content: {
+    height: "100%",
+  },
+});
 
 export default function CustomTabs({ tabValues }: TpropsArray) {
   const classes = useStyles();
