+++ client/src/pages/CatalogExplorer/CatalogExplorer.tsx
@@ -4,12 +4,10 @@ import CatalogCard from "../../components/CatalogCard/CatalogCard";
 //Component
 import CatalogCreator from "../../components/CatalogCard/CatalogCreator";
 // MUI
+import { makeStyles } from "@material-ui/core/styles";
 import { CircularProgress } from "@material-ui/core";
 
 import CustomAlert from "../../components/Alert/CustomAlert";
-
-//STYLES
-import useStyles from "./styles";
 
 type TcatalogCard = {
   id: string;
@@ -20,6 +18,23 @@ type TcatalogCard = {
   productCount: number;
   className?: any;
 };
+const useStyles = makeStyles((theme) => ({
+  gridContainer: {
+    display: "grid",
+    gridTemplateColumns: "repeat(auto-fit, minmax( 290px, 1fr))",
+    gap: "16px",
+  },
+  containerNoData: {
+    display: "grid",
+    gridTemplateColumns: "repeat(5, 1fr)",
+    gap: "16px",
+  },
+  loading: {
+    width: "100%",
+    display: "flex",
+    justifyContent: "center",
+  },
+}));
 
 const CatalogExplorer = () => {
   const classes = useStyles();
