+++ client/src/pages/DrawerAppbar/DrawerAppbar.tsx
@@ -20,10 +20,11 @@ import CatalogExplorer from "../CatalogExplorer/CatalogExplorer";
 import ProductsList from "../ProductsList/ProductsList";
 import ProductDetails from "../ProductDetails/ProductDetails";
 import AddProducts from "../../components/AddProducts/AddProducts";
+import "../../app.css";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
 import CustomNavBar from "../../components/CustomNavBar/CustomNavBar";
-import { useStyles } from "./styles";
+import { useStyles } from "./Styles";
 const drawerButtons = [
   {
     text: "Data Explorer",
@@ -85,7 +86,7 @@ export default function MiniDrawer() {
 
   return (
     <BrowserRouter>
-      <div className={classes.root}>
+      <div className={`${classes.root} App`}>
         <CustomNavBar
           className={clsx(classes.appBar, {
             [classes.appBarShift]: open,
