+++ client/src/pages/ProductDetails/ProductDetails.tsx
@@ -1,5 +1,6 @@
 import { Divider, Grid, CircularProgress } from "@material-ui/core";
-import { useEffect, useState } from "react";
+import { makeStyles, Theme } from "@material-ui/core/styles";
+import React, { useEffect, useState } from "react";
 import Accordion from "@material-ui/core/Accordion";
 import AccordionSummary from "@material-ui/core/AccordionSummary";
 import AccordionDetails from "@material-ui/core/AccordionDetails";
@@ -11,8 +12,6 @@ import getProductInfo from "../../api/getProductInfo";
 import { useQuery } from "@tanstack/react-query";
 import { useStore } from "../DrawerAppbar/DrawerAppbar";
 import CustomAlert from "../../components/Alert/CustomAlert";
-//STYLES
-import useStyles from "./styles";
 
 type Tproduct = {
   id: string;
@@ -25,6 +24,100 @@ type Tproduct = {
   allImages?: string[];
 };
 
+const useStyles = makeStyles((theme: Theme) => ({
+  gridContainer: {
+    background: theme.palette.background.paper,
+    height: "calc(100vh - 115px)",
+    width: "100%",
+    display: "flex",
+    justifyContent: "space-between",
+    gap: theme.spacing(4),
+    flexWrap: "nowrap",
+  },
+  leftBox: {
+    background: theme.palette.background.default,
+    height: "100%",
+    position: "relative",
+  },
+  rightBox: {
+    padding: theme.spacing(1, 2),
+    position: "relative",
+  },
+
+  imagesContainer: {
+    height: "inherit",
+  },
+  carousel: {
+    overflow: "auto",
+    height: "100%",
+    width: "100%",
+  },
+
+  img: {
+    cursor: "pointer",
+    width: "100%",
+    objectFit: "scale-down",
+    maxHeight: "170px",
+    paddingBottom: theme.spacing(0.5),
+    "&:hover": {
+      paddingBottom: theme.spacing(0),
+      borderBottom: "4px solid #5A67C5",
+    },
+  },
+  mainImgGrid: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    width: "100%",
+    height: "100%",
+    background: theme.palette.background.default,
+  },
+  mainImage: {
+    objectFit: "scale-down",
+    height: "100%",
+    width: "100%",
+    padding: theme.spacing(0.3),
+    flexGrow: 1,
+  },
+  accordionBox: {
+    width: "100%",
+  },
+  header: {
+    display: "grid",
+    gridTemplateColumns: "1fr",
+    gap: theme.spacing(2),
+  },
+  divider: {
+    marginBottom: theme.spacing(2),
+  },
+  details: {
+    display: "grid",
+    gridTemplateColumns: "1fr",
+  },
+  extraStyles: {
+    gap: theme.spacing(2),
+    padding: theme.spacing(0),
+    gridTemplateColumns: "2fr 3fr",
+    overflow: "hidden",
+  },
+  idBox: {
+    display: "flex",
+    justifyContent: "space-between",
+  },
+  productTitle: {
+    padding: theme.spacing(0, 0, 2, 0),
+  },
+  productDesc: {
+    display: "grid",
+    gap: theme.spacing(2),
+  },
+  center: {
+    display: "block",
+    position: "absolute",
+    top: "50%",
+    left: "50%",
+  },
+}));
 export default function ProductDetails() {
   const classes = useStyles();
   const { id: catalogId, productId } = useParams<{
@@ -105,7 +198,7 @@ export default function ProductDetails() {
 
           {isSuccess ? (
             <>
-              <div className={classes.headerDetails}>
+              <div className={classes.header}>
                 <Typography variant="subtitle1">Products Details</Typography>
                 <Typography variant="body2" className={classes.idBox}>
                   <Typography variant="caption">Product ID</Typography>
