+++ client/src/pages/ProductDetails/SummaryDetails.tsx
@@ -1,11 +1,50 @@
+import React from "react";
 import { Container, Typography, Toolbar, IconButton } from "@material-ui/core";
+import { makeStyles, Theme } from "@material-ui/core/styles";
 import CustomTabs from "../../components/CustomTabs";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faAngleRight } from "@fortawesome/free-solid-svg-icons";
 import DetailTable from "../../components/DetailTable";
 
-//STYLES
-import useStyles from "./styles";
+const useStyles = makeStyles((theme: Theme) => ({
+  container: {
+    display: "flex",
+    flexFlow: "column",
+    padding: 5,
+    paddingLeft: 8,
+    gap: 22,
+    lineHeight: "20px",
+    height: "100%",
+  },
+  header: {
+    display: "grid",
+    gridTemplateColumns: "1fr 3fr",
+    gap: 12,
+  },
+  image: {
+    width: "100%",
+  },
+  description: {
+    padding: 2,
+    lineHeight: "18px",
+  },
+  accordion: {
+    // border: "solid 1px black",
+  },
+  titleSection: {},
+  headerButtons: {
+    display: "grid",
+    gridTemplateColumns: "repeat(3, 1fr)",
+  },
+  currentButton: {
+    // borderBottom: "2px solid #6A5DF9",
+  },
+  toolbar: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "flex-start",
+  },
+}));
 
 export default function SummaryDetails({
   name,
